pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cartographer"
version = "1.0.0"
description = "Unified Archetypal Mapping Engine - Western Astrology + Human Design"
readme = "README.md"
authors = [
    {name = "Vibology", email = "joe@vibology.org"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn==0.27.1",
    "pydantic>=2.7.0",
    "kerykeion>=4.0.0",
    "pyswisseph>=2.10.3.2",
    "numpy==2.1.2",
    "pandas==2.2.3",
    "matplotlib",
    "svgpath2mpl",
    "Pillow",
    "geopy==2.4.1",
    "timezonefinder>=8.2.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "ipython",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["cartographer*"]

[tool.setuptools.package-data]
cartographer = [
    "data/*.json",
    "data/*.sqlite",
    "data/fonts/*.ttf",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
